Challenge with Tim from Udemy The first part of the challenge is to modify the show function in the Player class, so that it also prints the details for the weapon that the players have. -> checked
As well as printing the Player's name,
lives, level and score, it should also print the name of their weapon and the damage it inflicts. -> checked

Use the modified show function in main, to print out tim's details(including the weapon) instead of just the weapon.
So instead of just printing tim.weapon.name,
use the modified show function.

Hint : Review the video on String concatenation and interpolation, to remind yourself how to include
more complex expressions in a string.

-------------------------------------
For the second part of the challenge, modify the code in the main function, so that louise gets the sword before tim replaces it with a spear.

The point here is give Louise the same sword that Tim had, not to create a new one.

Use the show function to display the details for louise, to check that your code's working.